RootLookuper: fix CNAME handling
originally we were asking the same server to resolve the CNAME
target, which obviusly failed otherwise it would have been
included in the original response.

now we do a new request for the target name and the original
class/type, and append the result of the new request
to the original response.

Signed-off-by: Alejandro Mery <amery@jpi.io>